pyproject.toml
[tool.poetry]
name = "pricedisplay"
version = "0.2.6"
authors = ["Lumi"]
description = "Terminal display for the Finnish power price."
readme = "README.md"
license = "LICENSE.md"
repository = "https://github.com/YukiNeko-hime/pricedisplay"
classifiers = [
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
]

[project.urls]
"Source" = "https://github.com/YukiNeko-hime/pricedisplay"


[tool.poetry.dependencies]
python = "^3.9"
windows-curses = {version = "^2.3.1", platform = "win32"}
datetime = "^5.2"
usersettings = "^1.1.5"
pyyaml = "^6.0.1"
requests = "^2.31.0"
sparklines = "^0.4.2"
argparse = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
